move useRemoveSongsFromPlaylist()
