+++ projects/graphql-client/src/index.ts
@@ -27,3 +27,4 @@ export * from "./remote/mutations/useRevokeInvitation"
 export * from "./remote/mutations/useInviteToShare"
 export * from "./remote/mutations/useRenameShare"
 export * from "./remote/mutations/useRenamePlaylist"
+export * from "./remote/mutations/useRemoveSongsFromPlaylist"
